requirements.txt
fastapi
psycopg2
SQLAlchemy
FastAPI-SQLAlchemy
graphene 
graphene-sqlalchemy==3.0.0b3
starlette-graphene3
alembic
promise
pydantic
bcrypt
PyJWT
python-dotenv
redis
celery
flower
uvicorn[standard]